+++ packages/govern/public/index.ts
@@ -5,7 +5,7 @@ export { configure, ConfigurationObject } from './configure'
 export { dao, Dao } from './dao'
 export { daos, Daos } from './daos'
 export { query } from './query'
-export { createDao, CreateDaoOptions, CreateDaoParams } from './create-dao'
+export { createDao, CreateDaoOptions, CreateDaoParams } from './createDao'
 
 // Types
